% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obanalytics.R
\name{depth}
\alias{depth}
\title{Calculates and downloads depth changes from the OBADiah database}
\usage{
depth(con, start.time, end.time, exchange, pair, frequency = NULL, tz = "UTC")
}
\arguments{
\item{con}{a connection object as returned by \code{\link{connect}}}

\item{start.time}{POSIXct or character vector understood by \code{\link{ymd_hms}}}

\item{end.time}{POSIXct or character vector understood by \code{\link{ymd_hms}}}

\item{exchange}{a character vector with the name of the exchange}

\item{pair}{a character vector with the name of the pair}

\item{frequency}{if NULL, the actual depth changes are returned. Otherwise an integer number of seconds between depth samples.}

\item{tz}{a character vector with a time zone name understood by \code{\link{with_tz}} for the \code{timestamp} column in the output}
}
\value{
A data.table with one row per the depth change reported. The following information is provided for each depth change
\describe{
 \item{timestmap POSIXct}{timestamp of the depth change}
 \item{price numeric}{the price level for which the depth change is reported}
 \item{volume numeric}{The new amount available at the price level. Will be zero, when the last order with the given price will leave the order book.}
 \item{side character}{the side of the order book where the price level currently resides.  Either 'bid' or 'ask'.}
 \item{pair character}{a character vector with the pair name}
 \item{exchange character}{a character vector with the pair name}
}
}
\description{
A depth change is the new bid or ask volume  offered at some price in an exchange order book at some moment in time.
The change is caused by placement or cancellation of an order which happened on the exchange either at the time of the reported depth change
or between the current and pervious reported sample times if the depth changes are calculated using fixed frequency.
}
\details{
The depth changes are calculated as if before \code{start.time} the order book was empty, which effectively means that  the initial state of the order book at the start.time is conveyed.
The function ignores interruptions in the data and calculates depth changes over the interruptions. This may lead to the presence of long gaps between consecutive changes.
Not every order placement and/or cancellation produces a depth change. For example, if an order placement and cancellation are reported as happened at the same time, the depth change will
not be generated.
}
