% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obanalytics.R
\name{connect}
\alias{connect}
\title{Securely connects to the OBADiah database}
\usage{
connect(sslcert, sslkey, host, port, user = "obademo",
  dbname = "ob-analytics-prod")
}
\arguments{
\item{sslcert}{path to the client's SSL certificate file, signed by OBADiah database owner}

\item{sslkey}{path to the clients' SSL private key}

\item{host}{name or IP address of the PostgreSQL server running the OBADiah database}

\item{port}{port}

\item{user}{user}

\item{dbname}{name of PostgreSQL database}
}
\value{
the connection object as returned by DBI::dbConnect()
}
\description{
Establishes a secure TCP/IP connection to the OBADiah database and returns a connection object.
The object is used to communicate with the database.
}
