% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obanalytics.R
\name{connect}
\alias{connect}
\title{Securely connects to the OBADiah database and initializes an internal cache for the data}
\usage{
connect(
  host,
  port,
  sslcert = NULL,
  sslkey = NULL,
  sslrootcert = system.file("extdata/root.crt", package = packageName()),
  user = "obademo",
  dbname = "ob-analytics-prod",
  use.cache = TRUE
)
}
\arguments{
\item{host}{name or IP address of the PostgreSQL server running the OBADiah database}

\item{port}{port}

\item{sslcert}{path to the client's SSL certificate file, signed by OBADiah database owner}

\item{sslkey}{path to the clients' SSL private key}

\item{sslrootcert}{path to the clients' SSL root certificate}

\item{user}{user}

\item{dbname}{name of PostgreSQL database}

\item{use.cache}{if TRUE, the downloaded data will be cached on the client-side}
}
\value{
the connection object
}
\description{
Establishes a secure TCP/IP connection to the OBADiah database and returns a connection object.
The object is used to communicate with the database and to keep refereneces cached data
}
